FInal fixes for PIPE-11715 and some gcc warnings in PIPE-11838
git-svn-id: http://svnhq2.hq.eso.org/p2/trunk/Pipelines/cr2re/cr2rep@292313 318c8218-2dab-440d-9263-6f70bac63e5b

+++ cr2res/tests/cr2res_trace-test.c
@@ -497,7 +497,7 @@ static void test_cr2res_trace_get_height(void)
 
     cpl_test(res = cr2res_trace_get_height(trace, order_nb, trace_nb));
     //test output
-    cpl_test_eq(res, 174.1271552); // value analytically from test table
+    cpl_test_eq(res, (int) 174.1271552); // value analytically from test table
 
     //deallocate memory
     cpl_table_delete(trace);
